added proper cashflow

+++ backend/models/cashflow.go
@@ -1,10 +1,10 @@
 package models
 
 type CashflowSingle struct {
-	ID            int     `json:"id"`
+	ID            uint64  `json:"id"`
 	Date          string  `json:"date"`
 	TotalIncoming float64 `json:"totalIncoming"`
 	TotalOutgoing float64 `json:"totalOutgoing"`
 	TotalBalance  float64 `json:"totalBalance"`
-	CashbookID    int     `json:"cashbookId"`
+	CashbookID    uint64  `json:"cashbookId"`
 }
